refactor(matrices): move zeros, ones, scalarmul to RepMatrix
